fix: Disable system integrity protection (SIP) and Kext consent
This change removes the need for using a pre-patched NVRAM and cliclick.

The solution here is to provide a version of the UEFI shell (v2.2) lifted from
OpenCore 0.8.1 (https://github.com/acidanthera/OpenCorePkg/releases/tag/0.8.1).
Packer's cd_files command is used to create a bootable ISO of the UEFI shell.
Additionally, the VMX is patched to use bios.forceSetupOnce to boot into the
VMs firmware. From there the patched v2.2 of shell is used to perform a csrutil
like disable of system integrity protection (SIP). The system is then booted as
normal. Additionally, once booted to the installer terminal spctl kext-consent
disable is applied.